requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
sqlite3

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27
pillow==10.1.0

# AI/ML dependencies
openai==1.3.5
langchain==0.0.340
langchain-community==0.0.1
sentence-transformers==2.2.2
chromadb==0.4.17
numpy==1.24.4
pandas==2.1.3
scikit-learn==1.3.2

# HTTP requests and APIs
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# CORS
fastapi-cors==0.0.6

# File uploads
aiofiles==23.2.1

# Background tasks
celery==5.3.4
redis==5.0.1